Update tox configurations for Django 5.0 beta compatibility
Changes are made to the tox.ini file to include support for Django 5.0 beta version. Specifically, a new environment is added under the envlist section to test with Python 3.10 and Django 5.0 beta. Additionally, a new dependency line is added to specify the installation link for Django 5.0 beta. The existing environment versions for Django are commented out for now. This change will help in testing the project against the new Django version.